bugfix: drag and drop image on chat not working
