+++ src/frontend/src/modals/IOModal/components/chatView/chatInput/hooks/use-drag-and-drop.tsx
@@ -31,48 +31,11 @@ const useDragAndDrop = (
     setIsDragging(false);
   };
 
-  const onDrop = (e) => {
-    e.preventDefault();
-    if (e.dataTransfer.files && e.dataTransfer.files.length > 0) {
-      handleFiles(e.dataTransfer.files, setFiles, currentFlowId, setErrorData);
-      e.dataTransfer.clearData();
-    }
-    setIsDragging(false);
-  };
   return {
     dragOver,
     dragEnter,
     dragLeave,
-    onDrop,
   };
 };
 
-const handleFiles = (files, setFiles, currentFlowId, setErrorData) => {
-  if (files) {
-    const file = files?.[0];
-    const fileExtension = file.name.split(".").pop()?.toLowerCase();
-    if (
-      !fileExtension ||
-      !ALLOWED_IMAGE_INPUT_EXTENSIONS.includes(fileExtension)
-    ) {
-      console.log("Error uploading file");
-      setErrorData({
-        title: "Error uploading file",
-        list: [FS_ERROR_TEXT, SN_ERROR_TEXT],
-      });
-      return;
-    }
-    const uid = new ShortUniqueId();
-    const id = uid.randomUUID(3);
-    const type = files[0].type.split("/")[0];
-    const blob = files[0];
-
-    setFiles((prevFiles) => [
-      ...prevFiles,
-      { file: blob, loading: true, error: false, id, type },
-    ]);
-
-    // useFileUpload(blob, currentFlowId, setFiles, id);
-  }
-};
 export default useDragAndDrop;
